feat: add plot gen methods with logger support - feat: merge plot gen scripts into 'plotter.py` - chore: update `main.py` with lastest api call - fix: fix logger call across files - chore: add latest trajectory plots
