+++ main.py
@@ -2,31 +2,20 @@
 Main script for running the 6-body encounter simulation and visualization.
 """
 
-# %%
-# import
-from src.strw_amuse.config import PARAM_REF, PARAM_TEST
-from src.strw_amuse.gif_plotting import visualize_frames, visualize_initial_final_frames
+from src.strw_amuse.config import OUTPUT_DIR_LOGS, PARAM_REF, PARAM_TEST
 from src.strw_amuse.logging_config import setup_logging
+
+setup_logging(log_dir=OUTPUT_DIR_LOGS)
+
+from src.strw_amuse.plotter import plot_gif, plot_trajectory
 from src.strw_amuse.run_simulation import run_6_body_simulation
 
-# %%
-# init logger early
-setup_logging()
-
-# %%
-# run sim - no encounter
 frames_ref, outcome_ref = run_6_body_simulation(*PARAM_REF)
 
-# %%
-# visualize - no encounter -> no collisions
-visualize_initial_final_frames(frames=frames_ref, run_label="ref_case")
-visualize_frames(frames=frames_ref, run_label="ref_case")
+plot_gif(frames=frames_ref, run_label="ref_case")
+plot_trajectory(frames=frames_ref, run_label="ref_case")
 
-# %%
-# run sim - ionized
 frames_test, outcome_test = run_6_body_simulation(*PARAM_TEST)
 
-# %%
-# visualize - collisions -> creative ionized
-visualize_initial_final_frames(frames=frames_test, run_label="test_case")
-visualize_frames(frames=frames_test, run_label="test_case")
+plot_gif(frames=frames_test, run_label="test_case")
+plot_trajectory(frames=frames_test, run_label="test_case")
