+++ src/strw_amuse/logging_config.py
@@ -40,7 +40,7 @@ def _get_mpi_rank():
         return None
 
 
-def setup_logging(rank=None, log_dir=OUTPUT_DIR_LOGS, level=logging.INFO, name="strw_amuse"):
+def setup_logging(rank=None, log_dir=OUTPUT_DIR_LOGS, level=logging.INFO, name="src.strw_amuse"):
     """
     Configure logging for the package.
 
@@ -51,7 +51,7 @@ def setup_logging(rank=None, log_dir=OUTPUT_DIR_LOGS, level=logging.INFO, name="
     log_dir: str
         Directory where per-rank log files will be written if rank is not None.
     level: logging level
-    name: logger name
+    name: logger name (default `src.strw_amuse` to match package module names)
     """
     if rank is None:
         rank = _get_mpi_rank()
